+++ Templates/index.html
@@ -2,7 +2,7 @@
   <head>
     <meta charset="utf-8">
     <!-- Latest compiled and minified bootstrap CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
     <!-- minified slicknav css -->
     <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='css/slicknav.min.css') }}">
     <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='css/styles.css') }}">
